**** API CHANGE ****
RelaySpec - main constructor now takes the parameters in a sane order.
Removed the "reduced parameter" constructors, these are assymetric which is not in the spirit of this class - any client/server or otherwise asysymetric relationship should not be inferred at this level, especially with UDP (with the excpetion of example configs).